requirements.txt
pika
nats-py